requirements.txt
feedparser
google-generativeai
requests
python-dotenv
schedule
html2text
